Move common webpack code into a new package
This simplifies the process for making a new product and guards against
config drift.
