+++ packages/2017/package.json
@@ -45,6 +45,7 @@
     "babel-cli": "^6.24.1",
     "@hackoregon/civic-babel-presets": "^1.0.0-alpha.1c62c05f",
     "@hackoregon/dev-server": "^1.0.0-alpha.1c62c05f",
+    "@hackoregon/webpack-common": "^1.0.0-alpha.1c62c05f",
     "react-hot-loader": "^4.0.0"
   }
 }
